+++ plugin/web/extensions/sidebar/modals/datacatalog/components/content/YourDataPage/LocalDataTab.tsx
@@ -1,52 +1,35 @@
 import { CatalogItem, convertRaw } from "@web/extensions/sidebar/core/processCatalog";
-import { Select, Form } from "@web/sharedComponents";
+import { Form } from "@web/sharedComponents";
 import { InboxOutlined } from "@web/sharedComponents/Icon/icons";
-import Upload, { message, UploadProps, UploadFile } from "@web/sharedComponents/Upload";
-import { useState } from "react";
+import Upload, { UploadProps, UploadFile } from "@web/sharedComponents/Upload";
+import { RcFile } from "antd/lib/upload";
+import { useCallback, useState } from "react";
+
+import FileTypeSelect from "./FileTypeSelect";
 
 type Props = {
-  options: {
-    value: string;
-    label: string;
-  }[];
   onOpenDetails?: (data?: CatalogItem) => void;
 };
 
-const LocalDataTab: React.FC<Props> = ({ options, onOpenDetails }) => {
+const LocalDataTab: React.FC<Props> = ({ onOpenDetails }) => {
   const [fileList, setFileList] = useState<UploadFile[]>([]);
   const fileFormats = ".kml,.czml,.topojson,.geojson,.json,.gltf,.glb";
 
-  const props: UploadProps = {
-    name: "file",
-    multiple: false,
-    directory: false,
-    showUploadList: true,
-    accept: fileFormats,
-    listType: "picture",
-    onRemove: file => {
+  const onRemove = useCallback(
+    (file: UploadFile) => {
       const index = fileList.indexOf(file);
       const newFileList = fileList.slice();
       newFileList.splice(index, 1);
       setFileList(newFileList);
     },
-    onChange(info) {
-      const { status } = info.file;
-      if (status !== "uploading") {
-        console.log(info.file, info.fileList);
-      }
-      if (status === "done") {
-        message.success(`${info.file.name} file uploaded successfully.`);
-      } else if (status === "error") {
-        message.error(`${info.file.name} file upload failed.`);
-      }
-    },
-    onDrop(e) {
-      console.log("Dropped files", e.dataTransfer.files);
-    },
-    beforeUpload: (_file, files) => {
-      const filename = _file.name;
+    [fileList],
+  );
+
+  const beforeUpload = useCallback(
+    (file: RcFile, files: RcFile[]) => {
+      const filename = file.name;
       const id = "id" + Math.random().toString(16).slice(2);
-      const url = URL.createObjectURL(_file);
+      const url = URL.createObjectURL(file);
       const item = {
         id: id,
         description: "local data",
@@ -60,7 +43,7 @@ const LocalDataTab: React.FC<Props> = ({ options, onOpenDetails }) => {
       if (onOpenDetails) onOpenDetails(catalogItem);
 
       const reader = new FileReader();
-      reader.readAsText(_file);
+      reader.readAsText(file);
       let data;
       reader.onload = e => {
         data = e?.target?.result ?? "";
@@ -70,18 +53,25 @@ const LocalDataTab: React.FC<Props> = ({ options, onOpenDetails }) => {
       setFileList([...fileList, ...files]);
       return false;
     },
+    [fileList, onOpenDetails],
+  );
+
+  const props: UploadProps = {
+    name: "file",
+    multiple: false,
+    directory: false,
+    showUploadList: true,
+    accept: fileFormats,
+    listType: "picture",
+    onRemove: onRemove,
+    beforeUpload: beforeUpload,
     fileList,
   };
 
   return (
     <Form layout="vertical">
       <Form.Item name="file-type" label="Select file type">
-        <Select
-          defaultValue="auto"
-          style={{ width: "100%" }}
-          // onChange={handleChange}
-          options={options}
-        />
+        <FileTypeSelect />
       </Form.Item>
       <Form.Item label="Upload File">
         <Form.Item name="upload-file" style={{ height: 300, overflowY: "scroll" }}>
